create PostTypeRegistrar class
